fix: address review comments on node generations PR
- Clear liveGenerations and toolResults in clearResults to prevent memory
  growth across workflow closes
- Normalize non-record node_update results into { output: value } to
  keep the Generation.outputs contract stable
- Move assetToOutputValue from hook module to utils/nodeGenerations.ts
  so the pure utility stays framework-free (re-exported for compat)
- Pre-index edges by target in buildRunSubgraph for O(V+E) traversal
